meta-xilinx-mali400: skip sanity-meta-xilinx-mali400
Issue: LIN1024-4406

In fact, meta-xilinx-mali400 is added to layer.conf only when GPU
feature is enabled. But for now, it is added by default no mater GPU
feature is enabled or not. So we need to skip sanity-meta-xilinx-mali400
function to eliminate the annoying output:

    WARNING: You have included the meta-xilinx-mali400 layer,
    but 'libmali' has not been enabled in your DISTRO_FEATURES.
    Some bbappend files may not take effect. See the meta-xilinx-mali400
    README for details on enabling libmali - mali400 support.

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>